requirements.txt
boto3==1.9.199
certifi==2019.3.9
Click==7.0
Flask==1.0.3
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pbr==5.2.0
pydub==0.23.1
pymsql==0.9.3
python-dotenv==0.10.2
redis==3.2.1
rq==1.0
sentry-sdk==0.8.0
six==1.12.0
stevedore==1.30.1
urllib3==1.25.3
virtualenv==16.6.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
Werkzeug==0.15.4
uuid==1.30
